*Use GraphQLCodeRegistry to register data fetchers  - remove all deprecated datafetchers  - removed explicit PropertyDataFetcher declaration * split GraphQLJpaConvertertTest into jpa and GraphQL tests
