+++ graphql-jpa-query-autoconfigure/src/main/java/com/introproventures/graphql/jpa/query/autoconfigure/GraphQLShemaRegistrationImpl.java
@@ -3,12 +3,21 @@ package com.introproventures.graphql.jpa.query.autoconfigure;
 import java.util.LinkedHashSet;
 import java.util.Set;
 
+import graphql.schema.GraphQLCodeRegistry;
 import graphql.schema.GraphQLSchema;
 
 public class GraphQLShemaRegistrationImpl implements GraphQLShemaRegistration {
 
-	Set<GraphQLSchema> managedGraphQLSchemas = new LinkedHashSet<GraphQLSchema>();
+	private Set<GraphQLSchema> managedGraphQLSchemas = new LinkedHashSet<GraphQLSchema>();
 
+	private GraphQLCodeRegistry codeRegistry;
+
+    @Override
+    public void setCustomDataFetcher(GraphQLCodeRegistry codeRegistry) {
+        this.codeRegistry = codeRegistry;
+    }
+
+    @Override
 	public void register(GraphQLSchema graphQLSchema) {
 		managedGraphQLSchemas.add(graphQLSchema);
 	}
@@ -17,4 +26,7 @@ public class GraphQLShemaRegistrationImpl implements GraphQLShemaRegistration {
 		return managedGraphQLSchemas.toArray(new GraphQLSchema[] {});
 	}
 
+    public GraphQLCodeRegistry getCodeRegistry() {
+        return codeRegistry;
+    }
 }
