+++ graphql-jpa-query-autoconfigure/src/test/java/com/introproventures/graphql/jpa/query/autoconfigure/GraphQLSchemaAutoConfigurationTest.java
@@ -1,5 +1,7 @@
 package com.introproventures.graphql.jpa.query.autoconfigure;
 
+import static graphql.schema.FieldCoordinates.coordinates;
+import static graphql.schema.GraphQLCodeRegistry.newCodeRegistry;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import java.util.Map;
@@ -7,7 +9,6 @@ import java.util.Map;
 import graphql.GraphQL;
 import graphql.Scalars;
 import graphql.schema.DataFetcher;
-import graphql.schema.FieldCoordinates;
 import graphql.schema.GraphQLCodeRegistry;
 import graphql.schema.GraphQLFieldDefinition;
 import graphql.schema.GraphQLObjectType;
@@ -25,16 +26,19 @@ import org.springframework.test.context.junit4.SpringRunner;
 @RunWith(SpringRunner.class)
 @SpringBootTest(webEnvironment=WebEnvironment.NONE)
 public class GraphQLSchemaAutoConfigurationTest {
-    
+
     @Autowired
     private GraphQLSchema graphQLSchema;
 
     @Autowired
     private GraphQLJpaQueryProperties graphQLJpaQueryProperties;
-    
+
     @SpringBootApplication
     static class Application {
 
+        /**
+         * GraphQLSchemaFactoryBean maps all fields of the object "Mutation" to a new GraphQLObjectType
+         */
         @Component
         static class MutationGraphQLSchemaConfigurer implements GraphQLSchemaConfigurer {
 
@@ -42,54 +46,56 @@ public class GraphQLSchemaAutoConfigurationTest {
             public void configure(GraphQLShemaRegistration registry) {
 
                 GraphQLObjectType mutation = GraphQLObjectType.newObject()
-                        .name("mutation")
-                        .field(GraphQLFieldDefinition.newFieldDefinition()
-                                .name("greet")
-                                .type(Scalars.GraphQLString)
-                                .dataFetcher(environment -> {
-                                    return "hello world";
-                                }))
-                        .build();
-
-                GraphQLCodeRegistry codeRegistry = GraphQLCodeRegistry.newCodeRegistry()
-                    .dataFetcher(FieldCoordinates.coordinates("mutation", "greet"), (DataFetcher<String>) environment -> "hello world")
+                    .name("Mutation")
+                    .field(GraphQLFieldDefinition.newFieldDefinition()
+                        .name("greet")
+                        .type(Scalars.GraphQLString)
+                    )
                     .build();
+
+                GraphQLCodeRegistry codeRegistry = newCodeRegistry()
+                    .dataFetcher(
+                        coordinates("Mutation", "greet"),
+                        (DataFetcher<String>) environment -> "hello world"
+                    ).build();
 
                 GraphQLSchema graphQLSchema = GraphQLSchema.newSchema()
-                        .query(GraphQLObjectType.newObject().name("null"))
-                        .mutation(mutation)
-                        .codeRegistry(codeRegistry)
+                    .query(GraphQLObjectType.newObject().name("null"))
+                    .mutation(mutation)
+                    .codeRegistry(codeRegistry)
                     .build();
-                
+
                 registry.register(graphQLSchema);
             }
-        }        
+        }
+
+        /**
+         * GraphQLSchemaFactoryBean maps all fields of the object "Query" to a new GraphQLObjectType with the from GraphQLJpaQueryProperties
+         */
         @Component
         static class QueryGraphQLSchemaConfigurer implements GraphQLSchemaConfigurer {
 
             @Override
             public void configure(GraphQLShemaRegistration registry) {
+
                 GraphQLObjectType query = GraphQLObjectType.newObject()
-                        .name("query")
-                        .field(GraphQLFieldDefinition.newFieldDefinition()
-                                .name("hello")
-                                .type(Scalars.GraphQLString)
-                                .dataFetcher(environment -> {
-                                    return "world";
-                                }))
-                        .build();
+                    .name("Query")
+                    .field(GraphQLFieldDefinition.newFieldDefinition()
+                        .name("hello")
+                        .type(Scalars.GraphQLString))
+                    .build();
 
-              GraphQLCodeRegistry codeRegistry = GraphQLCodeRegistry.newCodeRegistry()
-                  .dataFetcher(FieldCoordinates.coordinates("query", "hello"), (DataFetcher<String>) environment -> "world")
-                  .build();
+                GraphQLSchema graphQLSchema = GraphQLSchema.newSchema()
+                    .query(query)
+                    .build();
 
-
-              GraphQLSchema graphQLSchema = GraphQLSchema.newSchema()
-                        .query(query)
-                        .codeRegistry(codeRegistry)
-                        .build();                
-                
                 registry.register(graphQLSchema);
+
+                GraphQLCodeRegistry codeRegistry = newCodeRegistry()
+                    .dataFetcher(coordinates("GraphQLBooks", "hello"), (DataFetcher<String>) environment -> "world")
+                    .build();
+
+                registry.setCustomDataFetcher(codeRegistry);
             }
         }
     }
@@ -102,17 +108,15 @@ public class GraphQLSchemaAutoConfigurationTest {
         // when
         Map<String, Object> result = graphQL.execute("query {hello}").getData();
         Map<String, Object> result2 = graphQL.execute("mutation {greet}").getData();
-        
+
         // then
         assertThat(result.toString()).isEqualTo("{hello=world}");
         assertThat(result2.toString()).isEqualTo("{greet=hello world}");
-        
+
         assertThat(graphQLSchema.getQueryType())
-                .extracting(GraphQLObjectType::getName, GraphQLObjectType::getDescription)
-                .containsExactly("GraphQLBooks", "GraphQL Books Schema Description");
-        	
+            .extracting(GraphQLObjectType::getName, GraphQLObjectType::getDescription)
+            .containsExactly("GraphQLBooks", "GraphQL Books Schema Description");
     }
-    
 
     @Test
     public void defaultConfigurationProperties() {
@@ -121,6 +125,5 @@ public class GraphQLSchemaAutoConfigurationTest {
         assertThat(graphQLJpaQueryProperties.isUseDistinctParameter()).isFalse();
         assertThat(graphQLJpaQueryProperties.isToManyDefaultOptional()).isTrue();
     }
-        
 
 }
