+++ graphql-jpa-query-schema/src/main/java/com/introproventures/graphql/jpa/query/schema/impl/GraphQLJpaSchemaBuilder.java
@@ -49,8 +49,9 @@ import com.introproventures.graphql.jpa.query.schema.impl.PredicateFilter.Criter
 import graphql.Assert;
 import graphql.Scalars;
 import graphql.schema.Coercing;
-import graphql.schema.DataFetcher;
+import graphql.schema.FieldCoordinates;
 import graphql.schema.GraphQLArgument;
+import graphql.schema.GraphQLCodeRegistry;
 import graphql.schema.GraphQLEnumType;
 import graphql.schema.GraphQLFieldDefinition;
 import graphql.schema.GraphQLInputObjectField;
@@ -63,18 +64,18 @@ import graphql.schema.GraphQLOutputType;
 import graphql.schema.GraphQLSchema;
 import graphql.schema.GraphQLType;
 import graphql.schema.GraphQLTypeReference;
-import graphql.schema.PropertyDataFetcher;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 /**
  * JPA specific schema builder implementation of {code #GraphQLSchemaBuilder} interface
- * 
+ *
  * @author Igor Dianov
  *
  */
 public class GraphQLJpaSchemaBuilder implements GraphQLSchemaBuilder {
+    private static final Logger log = LoggerFactory.getLogger(GraphQLJpaSchemaBuilder.class);
 
     private static final String AND = "AND";
     private static final String OR = "OR";
@@ -84,36 +85,37 @@ public class GraphQLJpaSchemaBuilder implements GraphQLSchemaBuilder {
 
     public static final String PAGE_START_PARAM_NAME = "start";
     public static final String PAGE_LIMIT_PARAM_NAME = "limit";
-    
+
     public static final String QUERY_SELECT_PARAM_NAME = "select";
     public static final String QUERY_WHERE_PARAM_NAME = "where";
     public static final String QUERY_LOGICAL_PARAM_NAME = "logical";
 
     public static final String SELECT_DISTINCT_PARAM_NAME = "distinct";
-    
+
     protected NamingStrategy namingStrategy = new NamingStrategy() {};
-    
+
     public static final String ORDER_BY_PARAM_NAME = "orderBy";
-    
+
     private Map<Class<?>, GraphQLOutputType> classCache = new HashMap<>();
     private Map<EntityType<?>, GraphQLObjectType> entityCache = new HashMap<>();
     private Map<ManagedType<?>, GraphQLInputObjectType> inputObjectCache = new HashMap<>();
     private Map<ManagedType<?>, GraphQLInputObjectType> subqueryInputObjectCache = new HashMap<>();
     private Map<Class<?>, GraphQLObjectType> embeddableOutputCache = new HashMap<>();
     private Map<Class<?>, GraphQLInputObjectType> embeddableInputCache = new HashMap<>();
-    
-    private static final Logger log = LoggerFactory.getLogger(GraphQLJpaSchemaBuilder.class);
+
+
+    private GraphQLCodeRegistry.Builder codeRegistry;
 
     private EntityManager entityManager;
-     
+
     private String name = "GraphQLJPAQuery";
-    
+
     private String description = "GraphQL Schema for all entities in this JPA application";
 
     private boolean isUseDistinctParameter = false;
     private boolean isDefaultDistinct = true;
     // the many end is a collection, and it is always optional by default (empty collection)
-    private boolean toManyDefaultOptional = true; 
+    private boolean toManyDefaultOptional = true;
 
     public GraphQLJpaSchemaBuilder(EntityManager entityManager) {
         this.entityManager = entityManager;
@@ -124,14 +126,17 @@ public class GraphQLJpaSchemaBuilder implements GraphQLSchemaBuilder {
      */
     @Override
     public GraphQLSchema build() {
+
+        codeRegistry = GraphQLCodeRegistry.newCodeRegistry();
+
         return GraphQLSchema.newSchema()
             .query(getQueryType())
+            .codeRegistry(getCodeRegistry())
             .build();
     }
 
     private GraphQLObjectType getQueryType() {
-        GraphQLObjectType.Builder queryType = 
-            GraphQLObjectType.newObject()
+        GraphQLObjectType.Builder queryType = GraphQLObjectType.newObject()
                 .name(this.name)
                 .description(this.description);
 
@@ -142,7 +147,7 @@ public class GraphQLJpaSchemaBuilder implements GraphQLSchemaBuilder {
                 .map(this::getQueryFieldByIdDefinition)
                 .collect(Collectors.toList())
         );
-        
+
         queryType.fields(
             entityManager.getMetamodel()
                 .getEntities().stream()
@@ -154,12 +159,27 @@ public class GraphQLJpaSchemaBuilder implements GraphQLSchemaBuilder {
         return queryType.build();
     }
 
+    private GraphQLCodeRegistry getCodeRegistry() {
+
+        for (EntityType<?> entity : entityManager.getMetamodel().getEntities()) {
+            if (!isNotIgnored(entity)) {
+                continue;
+            }
+            codeRegistry.dataFetcher(FieldCoordinates.coordinates(this.name, entity.getName()),
+                new GraphQLJpaSimpleDataFetcher(entityManager, entity, toManyDefaultOptional));
+
+            codeRegistry.dataFetcher(FieldCoordinates.coordinates(this.name, namingStrategy.pluralize(entity.getName())),
+                new GraphQLJpaQueryDataFetcher(entityManager, entity, isDefaultDistinct, toManyDefaultOptional));
+        }
+
+        return codeRegistry.build();
+    }
+
     private GraphQLFieldDefinition getQueryFieldByIdDefinition(EntityType<?> entityType) {
         return GraphQLFieldDefinition.newFieldDefinition()
                 .name(entityType.getName())
                 .description(getSchemaDescription(entityType))
                 .type(getObjectType(entityType))
-                .dataFetcher(new GraphQLJpaSimpleDataFetcher(entityManager, entityType, toManyDefaultOptional))
                 .arguments(entityType.getAttributes().stream()
                     .filter(this::isValidInput)
                     .filter(this::isNotIgnored)
@@ -168,10 +188,21 @@ public class GraphQLJpaSchemaBuilder implements GraphQLSchemaBuilder {
                     .collect(Collectors.toList())
                 )
                 .build();
-    }    
+    }
+
+    private void registerManyToManyDataFetcher(EntityType baseEntity, PluralAttribute<Object, Object, Object> attribute) {
+
+        codeRegistry.dataFetcher(FieldCoordinates.coordinates(baseEntity.getName(), attribute.getName()),
+            new GraphQLJpaOneToManyDataFetcher(entityManager,
+                baseEntity,
+                toManyDefaultOptional,
+                isDefaultDistinct,
+                attribute
+            ));
+    }
 
     private GraphQLFieldDefinition getQueryFieldSelectDefinition(EntityType<?> entityType) {
-        
+
         GraphQLObjectType pageType = GraphQLObjectType.newObject()
                 .name(namingStrategy.pluralize(entityType.getName()))
                 .description("Query response wrapper object for " + entityType.getName() + ".  When page is requested, this object will be returned with query metadata.")
@@ -202,10 +233,6 @@ public class GraphQLJpaSchemaBuilder implements GraphQLSchemaBuilder {
                     + "Use the '"+QUERY_SELECT_PARAM_NAME+"' field to request actual fields. "
                     + "Use the '"+ORDER_BY_PARAM_NAME+"' on a field to specify sort order for each field. ")
                 .type(pageType)
-                .dataFetcher(new GraphQLJpaQueryDataFetcher(entityManager, 
-                                                            entityType, 
-                                                            isDefaultDistinct, 
-                                                            toManyDefaultOptional))
                 .argument(paginationArgument)
                 .argument(getWhereArgument(entityType));
         if (isUseDistinctParameter) {
@@ -229,9 +256,9 @@ public class GraphQLJpaSchemaBuilder implements GraphQLSchemaBuilder {
     private GraphQLArgument getWhereArgument(ManagedType<?> managedType) {
         return whereArgumentsMap.computeIfAbsent(managedType.getJavaType(), (javaType) -> computeWhereArgument(managedType));
     }
-    
+
     private GraphQLArgument computeWhereArgument(ManagedType<?> managedType) {
-    	String type=resolveWhereArgumentTypeName(managedType);
+        String type=resolveWhereArgumentTypeName(managedType);
 
         GraphQLInputObjectType whereInputObject = GraphQLInputObjectType.newInputObject()
             .name(type)
@@ -252,18 +279,18 @@ public class GraphQLJpaSchemaBuilder implements GraphQLSchemaBuilder {
                                           .name("EXISTS")
                                           .description("Logical EXISTS subquery expression")
                                           .type(new GraphQLList(getSubqueryInputType(managedType)))
-                                          .build()            
+                                          .build()
             )
             .field(GraphQLInputObjectField.newInputObjectField()
                                           .name("NOT_EXISTS")
                                           .description("Logical NOT EXISTS subquery expression")
                                           .type(new GraphQLList(getSubqueryInputType(managedType)))
-                                          .build()            
+                                          .build()
             )
             .fields(managedType.getAttributes().stream()
                                                .filter(this::isValidInput)
                                                .filter(this::isNotIgnored)
-                                               .filter(this::isNotIgnoredFilter) 
+                                               .filter(this::isNotIgnoredFilter)
                                                .map(this::getWhereInputField)
                                                .collect(Collectors.toList())
             )
@@ -275,34 +302,34 @@ public class GraphQLJpaSchemaBuilder implements GraphQLSchemaBuilder {
                                                .collect(Collectors.toList())
             )
             .build();
-        
+
         return GraphQLArgument.newArgument()
                               .name(QUERY_WHERE_PARAM_NAME)
                               .description("Where logical specification")
                               .type(whereInputObject)
                               .build();
-        
+
     }
 
     private String resolveWhereArgumentTypeName(ManagedType<?> managedType) {
         String typeName=resolveTypeName(managedType);
-        
+
         return namingStrategy.pluralize(typeName)+"CriteriaExpression";
     }
-    
+
     private String resolveSubqueryArgumentTypeName(ManagedType<?> managedType) {
         String typeName=resolveTypeName(managedType);
-        
+
         return namingStrategy.pluralize(typeName)+"SubqueryCriteriaExpression";
     }
 
     private GraphQLInputObjectType getSubqueryInputType(ManagedType<?> managedType) {
         return subqueryInputObjectCache.computeIfAbsent(managedType, this::computeSubqueryInputType);
     }
-    
+
     private GraphQLInputObjectType computeSubqueryInputType(ManagedType<?> managedType) {
         String type=resolveSubqueryArgumentTypeName(managedType);
-        
+
          Builder whereInputObject = GraphQLInputObjectType.newInputObject()
             .name(type)
             .description("Where logical AND specification of the provided list of criteria expressions")
@@ -322,52 +349,52 @@ public class GraphQLJpaSchemaBuilder implements GraphQLSchemaBuilder {
                                           .name("EXISTS")
                                           .description("Logical EXISTS subquery expression")
                                           .type(new GraphQLList(new GraphQLTypeReference(type)))
-                                          .build()            
+                                          .build()
             )
             .field(GraphQLInputObjectField.newInputObjectField()
                                           .name("NOT_EXISTS")
                                           .description("Logical NOT EXISTS subquery expression")
                                           .type(new GraphQLList(new GraphQLTypeReference(type)))
-                                          .build()            
+                                          .build()
             )
             .fields(managedType.getAttributes().stream()
-                                               .filter(this::isValidAssociation)    
+                                               .filter(this::isValidAssociation)
                                                .filter(this::isNotIgnored)
                                                .filter(this::isNotIgnoredFilter)
                                                .map(this::getWhereInputRelationField)
                                                .collect(Collectors.toList())
             );
-        
+
         return whereInputObject.build();
-        
-    }     
-        
+
+    }
+
     private String resolveTypeName(ManagedType<?> managedType) {
         String typeName="";
-        
+
         if (managedType instanceof EmbeddableType){
             typeName = managedType.getJavaType().getSimpleName()+"EmbeddableType";
         } else if (managedType instanceof EntityType) {
             typeName = ((EntityType<?>)managedType).getName();
         }
-        
+
         return typeName;
     }
 
     private GraphQLInputObjectType getWhereInputType(ManagedType<?> managedType) {
         return inputObjectCache.computeIfAbsent(managedType, this::computeWhereInputType);
     }
-    
+
     private String resolveWhereInputTypeName(ManagedType<?> managedType) {
         String typeName=resolveTypeName(managedType);
 
         return namingStrategy.pluralize(typeName)+"RelationCriteriaExpression";
-        
+
     }
-    
+
     private GraphQLInputObjectType computeWhereInputType(ManagedType<?> managedType) {
         String type=resolveWhereInputTypeName(managedType);
-        
+
          Builder whereInputObject = GraphQLInputObjectType.newInputObject()
             .name(type)
             .description("Where logical AND specification of the provided list of criteria expressions")
@@ -387,13 +414,13 @@ public class GraphQLJpaSchemaBuilder implements GraphQLSchemaBuilder {
                                           .name("EXISTS")
                                           .description("Logical EXISTS subquery expression")
                                           .type(new GraphQLList(getSubqueryInputType(managedType)))
-                                          .build()            
+                                          .build()
             )
             .field(GraphQLInputObjectField.newInputObjectField()
                                           .name("NOT_EXISTS")
                                           .description("Logical NOT EXISTS subquery expression")
                                           .type(new GraphQLList(getSubqueryInputType(managedType)))
-                                          .build()            
+                                          .build()
             )
             .fields(managedType.getAttributes().stream()
                                                .filter(this::isValidInput)
@@ -403,21 +430,21 @@ public class GraphQLJpaSchemaBuilder implements GraphQLSchemaBuilder {
                                                .collect(Collectors.toList())
             )
             .fields(managedType.getAttributes().stream()
-                                               .filter(this::isValidAssociation)    
+                                               .filter(this::isValidAssociation)
                                                .filter(this::isNotIgnored)
                                                .filter(this::isNotIgnoredFilter)
                                                .map(this::getWhereInputRelationField)
                                                .collect(Collectors.toList())
             );
-         
-        
+
+
         return whereInputObject.build();
-        
-    }    
-    
+
+    }
+
     private GraphQLInputObjectField getWhereInputRelationField(Attribute<?,?> attribute) {
         ManagedType<?> foreignType = getForeignType(attribute);
-        
+
         String type = resolveWhereInputTypeName(foreignType);
         String description = getSchemaDescription(attribute);
 
@@ -425,9 +452,9 @@ public class GraphQLJpaSchemaBuilder implements GraphQLSchemaBuilder {
                                       .name(attribute.getName())
                                       .description(description)
                                       .type(new GraphQLTypeReference(type))
-                                      .build(); 
+                                      .build();
     }
-    
+
     private GraphQLInputObjectField getWhereInputField(Attribute<?,?> attribute) {
         GraphQLInputType type = getWhereAttributeType(attribute);
         String description = getSchemaDescription(attribute);
@@ -437,20 +464,20 @@ public class GraphQLJpaSchemaBuilder implements GraphQLSchemaBuilder {
                                           .name(attribute.getName())
                                           .description(description)
                                           .type(type)
-                                          .build(); 
+                                          .build();
         }
 
         throw new IllegalArgumentException("Attribute " + attribute.getName() + " cannot be mapped as an Input Argument");
     }
 
     private Map<String, GraphQLInputType> whereAttributesMap = new HashMap<>();
-    
+
     private GraphQLInputType getWhereAttributeType(Attribute<?,?> attribute) {
         String type =  namingStrategy.singularize(attribute.getName())+attribute.getDeclaringType().getJavaType().getSimpleName()+"Criteria";
 
         if(whereAttributesMap.containsKey(type))
            return whereAttributesMap.get(type);
-       
+
         GraphQLInputObjectType.Builder builder = GraphQLInputObjectType.newInputObject()
             .name(type)
             .description("Criteria expression specification of "+namingStrategy.singularize(attribute.getName())+" attribute in entity " + attribute.getDeclaringType().getJavaType())
@@ -477,8 +504,8 @@ public class GraphQLJpaSchemaBuilder implements GraphQLSchemaBuilder {
                 .description("Not Equals criteria")
                 .type(getAttributeInputType(attribute))
                 .build()
-            );
-       
+           );
+
             if(!attribute.getJavaType().isEnum()) {
                 if(!attribute.getJavaType().equals(String.class)) {
                     builder.field(GraphQLInputObjectField.newInputObjectField()
@@ -506,7 +533,7 @@ public class GraphQLJpaSchemaBuilder implements GraphQLSchemaBuilder {
                         .build()
                     );
                 }
-           
+
                 if(attribute.getJavaType().equals(String.class)) {
                     builder.field(GraphQLInputObjectField.newInputObjectField()
                         .name(Criteria.LIKE.name())
@@ -532,10 +559,10 @@ public class GraphQLJpaSchemaBuilder implements GraphQLSchemaBuilder {
                         .type(getAttributeInputType(attribute))
                         .build()
                     );
-                } 
-                else if (attribute.getJavaMember().getClass().isAssignableFrom(Field.class) 
+                }
+                else if (attribute.getJavaMember().getClass().isAssignableFrom(Field.class)
                         && Field.class.cast(attribute.getJavaMember())
-                                      .isAnnotationPresent(Convert.class)) 
+                                      .isAnnotationPresent(Convert.class))
                 {
                     builder.field(GraphQLInputObjectField.newInputObjectField()
                                                          .name(Criteria.LOCATE.name())
@@ -544,7 +571,7 @@ public class GraphQLJpaSchemaBuilder implements GraphQLSchemaBuilder {
                                                          .build());
                 }
             }
-            
+
             builder.field(GraphQLInputObjectField.newInputObjectField()
                 .name(Criteria.IS_NULL.name())
                 .description("Is Null criteria")
@@ -569,27 +596,27 @@ public class GraphQLJpaSchemaBuilder implements GraphQLSchemaBuilder {
                 .type(new GraphQLList(getAttributeInputType(attribute)))
                 .build()
             )
-           .field(GraphQLInputObjectField.newInputObjectField()
+            .field(GraphQLInputObjectField.newInputObjectField()
                 .name(Criteria.BETWEEN.name())
                 .description("Between criteria")
                 .type(new GraphQLList(getAttributeInputType(attribute)))
                 .build()
-           )
-           .field(GraphQLInputObjectField.newInputObjectField()
+            )
+            .field(GraphQLInputObjectField.newInputObjectField()
                 .name(Criteria.NOT_BETWEEN.name())
                 .description("Not Between criteria")
                 .type(new GraphQLList(getAttributeInputType(attribute)))
                 .build()
-           );
+            );
 
        GraphQLInputType answer = builder.build();
-       
+
        whereAttributesMap.putIfAbsent(type, answer);
-       
+
        return answer;
-       
+
     }
-    
+
     private GraphQLArgument getArgument(Attribute<?,?> attribute) {
         GraphQLInputType type = getAttributeInputType(attribute);
         String description = getSchemaDescription(attribute);
@@ -600,7 +627,7 @@ public class GraphQLJpaSchemaBuilder implements GraphQLSchemaBuilder {
                 .description(description)
                 .build();
     }
-    
+
     private GraphQLType getEmbeddableType(EmbeddableType<?> embeddableType, boolean input) {
         if (input && embeddableInputCache.containsKey(embeddableType.getJavaType()))
             return embeddableInputCache.get(embeddableType.getJavaType());
@@ -635,23 +662,23 @@ public class GraphQLJpaSchemaBuilder implements GraphQLSchemaBuilder {
         } else{
             embeddableOutputCache.putIfAbsent(embeddableType.getJavaType(), (GraphQLObjectType) graphQLType);
         }
-        
+
         return graphQLType;
     }
-    
+
 
     private GraphQLObjectType getObjectType(EntityType<?> entityType) {
         return entityCache.computeIfAbsent(entityType, this::computeObjectType);
     }
-    
-    
+
+
     private GraphQLObjectType computeObjectType(EntityType<?> entityType) {
-    	return GraphQLObjectType.newObject()
-				                .name(entityType.getName())
-				                .description(getSchemaDescription(entityType))
-				                .fields(getEntityAttributesFields(entityType))
-				                .fields(getTransientFields(entityType))
-				                .build();
+        return GraphQLObjectType.newObject()
+                            .name(entityType.getName())
+                            .description(getSchemaDescription(entityType))
+                            .fields(getEntityAttributesFields(entityType))
+                            .fields(getTransientFields(entityType))
+                            .build();
     }
 
     private List<GraphQLFieldDefinition> getEntityAttributesFields(EntityType<?> entityType) {
@@ -671,19 +698,17 @@ public class GraphQLJpaSchemaBuilder implements GraphQLSchemaBuilder {
                                   .map(this::getJavaFieldDefinition)
                                  .collect(Collectors.toList());
     }
-    
+
     @SuppressWarnings( { "rawtypes" } )
     private GraphQLFieldDefinition getJavaFieldDefinition(AttributePropertyDescriptor propertyDescriptor) {
-    	GraphQLOutputType type = getGraphQLTypeFromJavaType(propertyDescriptor.getPropertyType());
-        DataFetcher dataFetcher = PropertyDataFetcher.fetching(propertyDescriptor.getName());
-        
+       GraphQLOutputType type = getGraphQLTypeFromJavaType(propertyDescriptor.getPropertyType());
+
         String description = propertyDescriptor.getSchemaDescription().orElse(null);
 
         return GraphQLFieldDefinition.newFieldDefinition()
                                      .name(propertyDescriptor.getName())
                                      .description(description)
                                      .type(type)
-                                     .dataFetcher(dataFetcher)
                                      .build();
     }
 
@@ -696,7 +721,6 @@ public class GraphQLJpaSchemaBuilder implements GraphQLSchemaBuilder {
         GraphQLOutputType type = getAttributeOutputType(attribute);
 
         List<GraphQLArgument> arguments = new ArrayList<>();
-        DataFetcher dataFetcher = PropertyDataFetcher.fetching(attribute.getName());
 
         // Only add the orderBy argument for basic attribute types
         if (attribute instanceof SingularAttribute
@@ -711,14 +735,14 @@ public class GraphQLJpaSchemaBuilder implements GraphQLSchemaBuilder {
         }
 
         // Get the fields that can be queried on (i.e. Simple Types, no Sub-Objects)
-        if (attribute instanceof SingularAttribute 
+        if (attribute instanceof SingularAttribute
             && attribute.getPersistentAttributeType() != Attribute.PersistentAttributeType.BASIC) {
             ManagedType foreignType = getForeignType(attribute);
             SingularAttribute<?,?> singularAttribute = SingularAttribute.class.cast(attribute);
 
             // TODO fix page count query
             arguments.add(getWhereArgument(foreignType));
-            
+
             // to-one end could be optional  
             arguments.add(optionalArgument(singularAttribute.isOptional()));
 
@@ -726,32 +750,27 @@ public class GraphQLJpaSchemaBuilder implements GraphQLSchemaBuilder {
         else if (attribute instanceof PluralAttribute
             && (attribute.getPersistentAttributeType() == Attribute.PersistentAttributeType.ONE_TO_MANY
                 || attribute.getPersistentAttributeType() == Attribute.PersistentAttributeType.MANY_TO_MANY)) {
-            Assert.assertNotNull(baseEntity, "For attribute "+attribute.getName() + " cannot find declaring type!");
+            Assert.assertNotNull(baseEntity, "For attribute " + attribute.getName() + " cannot find declaring type!");
             EntityType elementType =  (EntityType) ((PluralAttribute) attribute).getElementType();
 
             arguments.add(getWhereArgument(elementType));
-            
+
             // make it configurable via builder api
             arguments.add(optionalArgument(toManyDefaultOptional));
 
             if (attribute.getPersistentAttributeType() == Attribute.PersistentAttributeType.MANY_TO_MANY) {
-                dataFetcher = new GraphQLJpaOneToManyDataFetcher(entityManager,
-                                                                 baseEntity,
-                                                                 toManyDefaultOptional,
-                                                                 isDefaultDistinct,
-                                                                 (PluralAttribute) attribute);
+                registerManyToManyDataFetcher(baseEntity, (PluralAttribute) attribute);
             }
         }
-        
+
         return GraphQLFieldDefinition.newFieldDefinition()
                 .name(attribute.getName())
                 .description(getSchemaDescription(attribute))
                 .type(type)
-                .dataFetcher(dataFetcher)
                 .arguments(arguments)
                 .build();
     }
-    
+
     private GraphQLArgument optionalArgument(Boolean defaultValue) {
         return GraphQLArgument.newArgument()
                 .name("optional")
@@ -760,14 +779,14 @@ public class GraphQLJpaSchemaBuilder implements GraphQLSchemaBuilder {
                 .defaultValue(defaultValue)
                 .build();
     }
-    
+
     protected ManagedType<?> getForeignType(Attribute<?,?> attribute) {
         if(SingularAttribute.class.isInstance(attribute))
             return (ManagedType<?>) ((SingularAttribute<?,?>) attribute).getType();
         else
             return (EntityType<?>) ((PluralAttribute<?, ?, ?>) attribute).getElementType();
     }
-    
+
     @SuppressWarnings( { "rawtypes" } )
     private GraphQLInputObjectField getInputObjectField(Attribute attribute) {
         GraphQLInputType type = getAttributeInputType(attribute);
@@ -784,7 +803,7 @@ public class GraphQLJpaSchemaBuilder implements GraphQLSchemaBuilder {
     }
 
     private GraphQLInputType getAttributeInputType(Attribute<?,?> attribute) {
-        
+
         try {
             return (GraphQLInputType) getAttributeType(attribute, true);
         } catch (ClassCastException e){
@@ -804,27 +823,27 @@ public class GraphQLJpaSchemaBuilder implements GraphQLSchemaBuilder {
     private GraphQLType getAttributeType(Attribute<?,?> attribute, boolean input) {
 
         if (isBasic(attribute)) {
-        	return getGraphQLTypeFromJavaType(attribute.getJavaType());
-        } 
+            return getGraphQLTypeFromJavaType(attribute.getJavaType());
+        }
         else if (isEmbeddable(attribute)) {
-        	EmbeddableType embeddableType = (EmbeddableType) ((SingularAttribute) attribute).getType();
-        	return getEmbeddableType(embeddableType, input);
-        } 
+            EmbeddableType embeddableType = (EmbeddableType) ((SingularAttribute) attribute).getType();
+            return getEmbeddableType(embeddableType, input);
+        }
         else if (isToMany(attribute)) {
             EntityType foreignType = (EntityType) ((PluralAttribute) attribute).getElementType();
-            
+
             return input ? getWhereInputType(foreignType) : new GraphQLList(new GraphQLTypeReference(foreignType.getName()));
-        } 
+        }
         else if (isToOne(attribute)) {
             EntityType foreignType = (EntityType) ((SingularAttribute) attribute).getType();
-            
+
             return input ? getWhereInputType(foreignType) : new GraphQLTypeReference(foreignType.getName());
-        } 
+        }
         else if (isElementCollection(attribute)) {
             Type foreignType = ((PluralAttribute) attribute).getElementType();
-            
+
             if(foreignType.getPersistenceType() == Type.PersistenceType.BASIC) {
-            	return new GraphQLList(getGraphQLTypeFromJavaType(foreignType.getJavaType()));
+                return new GraphQLList(getGraphQLTypeFromJavaType(foreignType.getJavaType()));
             }
         }
 
@@ -836,31 +855,31 @@ public class GraphQLJpaSchemaBuilder implements GraphQLSchemaBuilder {
     }
 
     protected final boolean isEmbeddable(Attribute<?,?> attribute) {
-    	return attribute.getPersistentAttributeType() == Attribute.PersistentAttributeType.EMBEDDED;
+        return attribute.getPersistentAttributeType() == Attribute.PersistentAttributeType.EMBEDDED;
     }
-    
+
     protected final boolean isBasic(Attribute<?,?> attribute) {
-    	return attribute.getPersistentAttributeType() == Attribute.PersistentAttributeType.BASIC;
+        return attribute.getPersistentAttributeType() == Attribute.PersistentAttributeType.BASIC;
     }
-    
+
     protected final boolean isElementCollection(Attribute<?,?> attribute) {
-    	return  attribute.getPersistentAttributeType() == Attribute.PersistentAttributeType.ELEMENT_COLLECTION;
+        return  attribute.getPersistentAttributeType() == Attribute.PersistentAttributeType.ELEMENT_COLLECTION;
     }
-    
+
     protected final boolean isToMany(Attribute<?,?> attribute) {
-    	return attribute.getPersistentAttributeType() == Attribute.PersistentAttributeType.ONE_TO_MANY
-        		|| attribute.getPersistentAttributeType() == Attribute.PersistentAttributeType.MANY_TO_MANY;
+        return attribute.getPersistentAttributeType() == Attribute.PersistentAttributeType.ONE_TO_MANY
+                || attribute.getPersistentAttributeType() == Attribute.PersistentAttributeType.MANY_TO_MANY;
     }
 
     protected final boolean isOneToMany(Attribute<?,?> attribute) {
         return attribute.getPersistentAttributeType() == Attribute.PersistentAttributeType.ONE_TO_MANY;
     }
-    
+
     protected final boolean isToOne(Attribute<?,?> attribute) {
-    	return attribute.getPersistentAttributeType() == Attribute.PersistentAttributeType.MANY_TO_ONE
-        		|| attribute.getPersistentAttributeType() == Attribute.PersistentAttributeType.ONE_TO_ONE;
+        return attribute.getPersistentAttributeType() == Attribute.PersistentAttributeType.MANY_TO_ONE
+                || attribute.getPersistentAttributeType() == Attribute.PersistentAttributeType.ONE_TO_ONE;
     }
-    
+
 
     protected final boolean isValidInput(Attribute<?,?> attribute) {
         return attribute.getPersistentAttributeType() == Attribute.PersistentAttributeType.BASIC ||
@@ -878,7 +897,7 @@ public class GraphQLJpaSchemaBuilder implements GraphQLSchemaBuilder {
                                  .getSchemaDescription(attribute.getName())
                                  .orElse(null);
     }
-    
+
     private String getSchemaDescription(EntityType<?> entityType) {
         return EntityIntrospector.introspect(entityType)
                                  .getSchemaDescription()
@@ -890,11 +909,11 @@ public class GraphQLJpaSchemaBuilder implements GraphQLSchemaBuilder {
                                  .getSchemaDescription()
                                  .orElse(null);
     }
-    
+
     private boolean isNotIgnored(EmbeddableType<?> attribute) {
         return isNotIgnored(attribute.getJavaType());
     }
-    
+
     private boolean isNotIgnored(Attribute<?,?> attribute) {
         return isNotIgnored(attribute.getJavaMember()) && isNotIgnored(attribute.getJavaType());
     }
@@ -902,7 +921,7 @@ public class GraphQLJpaSchemaBuilder implements GraphQLSchemaBuilder {
     private boolean isIdentity(Attribute<?,?> attribute) {
         return attribute instanceof SingularAttribute && ((SingularAttribute<?,?>)attribute).isId();
     }
-    
+
     private boolean isNotIgnored(EntityType<?> entityType) {
         return isNotIgnored(entityType.getJavaType());
     }
@@ -945,14 +964,14 @@ public class GraphQLJpaSchemaBuilder implements GraphQLSchemaBuilder {
         return false;
     }
 
-    
+
     @SuppressWarnings( "unchecked" )
     private GraphQLOutputType getGraphQLTypeFromJavaType(Class<?> clazz) {
         if (clazz.isEnum()) {
-            
+
             if (classCache.containsKey(clazz))
                 return classCache.get(clazz);
-            
+
             GraphQLEnumType.Builder enumBuilder = GraphQLEnumType.newEnum().name(clazz.getSimpleName());
             int ordinal = 0;
             for (Enum<?> enumValue : ((Class<Enum<?>>)clazz).getEnumConstants())
@@ -962,7 +981,7 @@ public class GraphQLJpaSchemaBuilder implements GraphQLSchemaBuilder {
             setNoOpCoercing(enumType);
 
             classCache.putIfAbsent(clazz, enumType);
-            
+
             return enumType;
         } else if (clazz.isArray()) {
             return GraphQLList.list(JavaScalars.of(clazz.getComponentType()));
@@ -972,21 +991,21 @@ public class GraphQLJpaSchemaBuilder implements GraphQLSchemaBuilder {
     }
 
     protected GraphQLInputType getFieldsEnumType(EntityType<?> entityType) {
-            
+
         GraphQLEnumType.Builder enumBuilder = GraphQLEnumType.newEnum().name(entityType.getName()+"FieldsEnum");
         final AtomicInteger ordinal = new AtomicInteger();
-        
+
         entityType.getAttributes().stream()
             .filter(this::isValidInput)
             .filter(this::isNotIgnored)
             .forEach(it -> enumBuilder.value(it.getName(), ordinal.incrementAndGet()));
-        
+
         GraphQLInputType answer = enumBuilder.build();
         setNoOpCoercing(answer);
 
         return answer;
     }
-    
+
     /**
      * JPA will deserialize Enum's for us...we don't want GraphQL doing it.
      *
@@ -1030,7 +1049,7 @@ public class GraphQLJpaSchemaBuilder implements GraphQLSchemaBuilder {
                     .value("DESC", "DESC", "Descending")
                     .build();
 
-    
+
     /**
      * @return the name
      */
@@ -1044,7 +1063,7 @@ public class GraphQLJpaSchemaBuilder implements GraphQLSchemaBuilder {
     @Override
     public GraphQLJpaSchemaBuilder name(String name) {
         this.name = name;
-        
+
         return this;
     }
 
@@ -1092,7 +1111,7 @@ public class GraphQLJpaSchemaBuilder implements GraphQLSchemaBuilder {
         this.namingStrategy = namingStrategy;
     }
 
-    
+
     static class NoOpCoercing implements Coercing<Object, Object> {
 
         @Override
@@ -1122,20 +1141,20 @@ public class GraphQLJpaSchemaBuilder implements GraphQLSchemaBuilder {
     @Override
     public GraphQLSchemaBuilder namingStrategy(NamingStrategy instance) {
         Assert.assertNotNull(instance, "instance is null");
-        
+
         this.namingStrategy = instance;
-        
+
         return this;
     }
 
-    
+
     public boolean isToManyDefaultOptional() {
         return toManyDefaultOptional;
     }
 
-    
+
     public void setToManyDefaultOptional(boolean toManyDefaultOptional) {
         this.toManyDefaultOptional = toManyDefaultOptional;
     }
-    
+
 }
