+++ graphql-jpa-query-schema/src/test/java/com/introproventures/graphql/jpa/query/converter/GraphQLJpaConverterTests.java
@@ -20,9 +20,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 import java.util.Collections;
 import java.util.List;
-import java.util.Map;
 
-import javax.persistence.EntityManager;
 import javax.persistence.Query;
 import javax.persistence.criteria.CriteriaBuilder;
 import javax.persistence.criteria.CriteriaQuery;
@@ -38,54 +36,12 @@ import com.introproventures.graphql.jpa.query.converter.model.JsonEntity;
 import com.introproventures.graphql.jpa.query.converter.model.TaskEntity;
 import com.introproventures.graphql.jpa.query.converter.model.TaskVariableEntity;
 import com.introproventures.graphql.jpa.query.converter.model.VariableValue;
-import com.introproventures.graphql.jpa.query.schema.GraphQLExecutor;
-import com.introproventures.graphql.jpa.query.schema.GraphQLSchemaBuilder;
-import com.introproventures.graphql.jpa.query.schema.impl.GraphQLJpaExecutor;
-import com.introproventures.graphql.jpa.query.schema.impl.GraphQLJpaSchemaBuilder;
+
 import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.test.context.SpringBootTest.WebEnvironment;
-import org.springframework.context.annotation.Bean;
-import org.springframework.test.context.TestPropertySource;
-import org.springframework.test.context.junit4.SpringRunner;
 
 
-@RunWith(SpringRunner.class)
-@SpringBootTest(webEnvironment=WebEnvironment.NONE, 
-                properties = "spring.datasource.data=GraphQLJpaConverterTests.sql")
-@TestPropertySource({"classpath:hibernate.properties"})
-public class GraphQLJpaConverterTests {
-    
-    @SpringBootApplication
-    static class Application {
-        @Bean
-        public GraphQLExecutor graphQLExecutor(final GraphQLSchemaBuilder graphQLSchemaBuilder) {
-            return new GraphQLJpaExecutor(graphQLSchemaBuilder.build());
-        }
+public class GraphQLJpaConverterTests extends BaseConverterTest {
 
-        @Bean
-        public GraphQLSchemaBuilder graphQLSchemaBuilder(final EntityManager entityManager) {
-            return new GraphQLJpaSchemaBuilder(entityManager)
-                .name("CustomAttributeConverterSchema")
-                .description("Custom Attribute Converter Schema");
-        }
-        
-    }
-    
-    @Autowired
-    private GraphQLExecutor executor;
-    
-    @Autowired
-    private EntityManager entityManager;
-    
-    @Test
-    public void contextLoads() {
-        
-    }
-    
     @Test
     @Transactional
     public void queryTester() {
@@ -126,10 +82,8 @@ public class GraphQLJpaConverterTests {
         CriteriaBuilder cb = entityManager.getCriteriaBuilder();
         CriteriaQuery<TaskVariableEntity> criteria = cb.createQuery(TaskVariableEntity.class);
         Root<TaskVariableEntity> taskVariable = criteria.from(TaskVariableEntity.class);
-        
-        Boolean object = Boolean.TRUE;
 
-        VariableValue<Boolean> variableValue = new VariableValue<>(object);
+        VariableValue<Boolean> variableValue = new VariableValue<>(Boolean.TRUE);
         criteria.select(taskVariable)
                 .where(cb.and(cb.equal(cb.lower(taskVariable.get("name")), "variable2")),
                               cb.in(taskVariable.get("value")).value(variableValue));
@@ -150,8 +104,8 @@ public class GraphQLJpaConverterTests {
         CriteriaQuery<TaskEntity> tasksQuery = cb.createQuery(TaskEntity.class);
         Root<TaskEntity> task = tasksQuery.from(TaskEntity.class);
         
-        Subquery<TaskVariableEntity> subquery = tasksQuery.subquery(TaskVariableEntity.class);
-        Root<TaskVariableEntity> taskVariables = subquery.from(TaskVariableEntity.class);
+        Subquery<TaskVariableEntity> subQuery = tasksQuery.subquery(TaskVariableEntity.class);
+        Root<TaskVariableEntity> taskVariables = subQuery.from(TaskVariableEntity.class);
         
         Predicate isOwner = cb.in(taskVariables.get("task")).value(task);
 
@@ -159,11 +113,11 @@ public class GraphQLJpaConverterTests {
                                 cb.equal(taskVariables.get("value"), new VariableValue<>(Boolean.TRUE)));
 
         Predicate var2 = cb.and(cb.equal(taskVariables.get("name"), "variable1"), 
-                                cb.equal(taskVariables.get("value"), new VariableValue<>(new String("data"))));
+                                cb.equal(taskVariables.get("value"), new VariableValue<>("data")));
         
         tasksQuery.select(task)
-                  .where(cb.and(cb.exists(subquery.select(taskVariables).where(cb.and(isOwner, var1))),
-                                cb.exists(subquery.select(taskVariables).where(cb.and(isOwner, var2)))));
+                  .where(cb.and(cb.exists(subQuery.select(taskVariables).where(cb.and(isOwner, var1))),
+                                cb.exists(subQuery.select(taskVariables).where(cb.and(isOwner, var2)))));
         // when:
         List<TaskEntity> result = entityManager.createQuery(tasksQuery).getResultList();
 
@@ -180,8 +134,8 @@ public class GraphQLJpaConverterTests {
         CriteriaQuery<TaskEntity> tasksQuery = cb.createQuery(TaskEntity.class);
         Root<TaskEntity> task = tasksQuery.from(TaskEntity.class);
         
-        Subquery<TaskVariableEntity> subquery = tasksQuery.subquery(TaskVariableEntity.class);
-        Root<TaskEntity> taskCorrelation = subquery.correlate(task);
+        Subquery<TaskVariableEntity> subQuery = tasksQuery.subquery(TaskVariableEntity.class);
+        Root<TaskEntity> taskCorrelation = subQuery.correlate(task);
         
         Join<TaskEntity, TaskVariableEntity> taskVariables = taskCorrelation.join("variables");
         
@@ -189,11 +143,11 @@ public class GraphQLJpaConverterTests {
                                 cb.equal(taskVariables.get("value"), new VariableValue<>(Boolean.TRUE)));
 
         Predicate var2 = cb.and(cb.equal(taskVariables.get("name"), "variable1"), 
-                                cb.equal(taskVariables.get("value"), new VariableValue<>(new String("data"))));
+                                cb.equal(taskVariables.get("value"), new VariableValue<>("data")));
         
         tasksQuery.select(task)
-                  .where(cb.and(cb.exists(subquery.select(taskVariables).where(var1)),
-                                cb.exists(subquery.select(taskVariables).where(var2))));
+                  .where(cb.and(cb.exists(subQuery.select(taskVariables).where(var1)),
+                                cb.exists(subQuery.select(taskVariables).where(var2))));
         // when:
         List<TaskEntity> result = entityManager.createQuery(tasksQuery).getResultList();
 
@@ -210,8 +164,8 @@ public class GraphQLJpaConverterTests {
         CriteriaQuery<TaskEntity> tasksQuery = cb.createQuery(TaskEntity.class);
         Root<TaskEntity> task = tasksQuery.from(TaskEntity.class);
         
-        Subquery<TaskVariableEntity> subquery = tasksQuery.subquery(TaskVariableEntity.class);
-        Root<TaskEntity> taskCorrelation = subquery.correlate(task);
+        Subquery<TaskVariableEntity> subQuery = tasksQuery.subquery(TaskVariableEntity.class);
+        Root<TaskEntity> taskCorrelation = subQuery.correlate(task);
         
         Join<TaskEntity, TaskVariableEntity> taskVariables = taskCorrelation.join("variables");
         
@@ -219,11 +173,11 @@ public class GraphQLJpaConverterTests {
                                 cb.equal(taskVariables.get("value"), new VariableValue<>(Boolean.TRUE)));
 
         Predicate var2 = cb.and(cb.equal(taskVariables.get("name"), "variable1"), 
-                                cb.equal(taskVariables.get("value"), new VariableValue<>(new String("data"))));
+                                cb.equal(taskVariables.get("value"), new VariableValue<>("data")));
         
         tasksQuery.select(task)
-                  .where(cb.and(cb.not(cb.exists(subquery.select(taskVariables).where(var1))),
-                                cb.not(cb.exists(subquery.select(taskVariables).where(var2)))
+                  .where(cb.and(cb.not(cb.exists(subQuery.select(taskVariables).where(var1))),
+                                cb.not(cb.exists(subQuery.select(taskVariables).where(var2)))
                                 )
                          );
         // when:
@@ -242,8 +196,8 @@ public class GraphQLJpaConverterTests {
         CriteriaQuery<TaskEntity> tasksQuery = cb.createQuery(TaskEntity.class);
         Root<TaskEntity> task = tasksQuery.from(TaskEntity.class);
         
-        Subquery<TaskVariableEntity> subquery = tasksQuery.subquery(TaskVariableEntity.class);
-        Root<TaskEntity> taskCorrelation = subquery.correlate(task);
+        Subquery<TaskVariableEntity> subQuery = tasksQuery.subquery(TaskVariableEntity.class);
+        Root<TaskEntity> taskCorrelation = subQuery.correlate(task);
         
         Join<TaskEntity, TaskVariableEntity> taskVariables = taskCorrelation.join("variables");
         
@@ -251,11 +205,11 @@ public class GraphQLJpaConverterTests {
                                 cb.equal(taskVariables.get("value"), new VariableValue<>(Boolean.TRUE)));
 
         Predicate var2 = cb.and(cb.equal(taskVariables.get("name"), "variable1"), 
-                                cb.equal(taskVariables.get("value"), new VariableValue<>(new String("data"))));
+                                cb.equal(taskVariables.get("value"), new VariableValue<>("data")));
         
         tasksQuery.select(task)
-                  .where(cb.not(cb.or(cb.exists(subquery.select(taskVariables).where(var1)),
-                                      cb.exists(subquery.select(taskVariables).where(var2)))
+                  .where(cb.not(cb.or(cb.exists(subQuery.select(taskVariables).where(var1)),
+                                      cb.exists(subQuery.select(taskVariables).where(var2)))
                                 )
                          );
         // when:
@@ -293,7 +247,7 @@ public class GraphQLJpaConverterTests {
         Root<JsonEntity> json = criteria.from(JsonEntity.class);
 
         criteria.select(json)
-                .where(builder.gt(builder.locate(json.<String>get("attributes"),"key"), 0));
+                .where(builder.gt(builder.locate(json.get("attributes"),"key"), 0));
         
         // when:
         List<?> result = entityManager.createQuery(criteria).getResultList();
@@ -302,533 +256,4 @@ public class GraphQLJpaConverterTests {
         assertThat(result).isNotEmpty();
         assertThat(result).hasSize(1);
     }
-    
-    @Test
-    public void queryJsonEntity() {
-        //given
-        String query = "query {" + 
-                "  JsonEntities {" + 
-                "    select {" + 
-                "      id" + 
-                "      firstName" + 
-                "      lastName" + 
-                "      attributes" + 
-                "    }" + 
-                "  }" + 
-                "}";
-        
-        String expected = "{JsonEntities={select=["
-                + "{id=1, firstName=john, lastName=doe, attributes={\"attr\":{\"key\":[\"1\",\"2\",\"3\",\"4\",\"5\"]}}}, "
-                + "{id=2, firstName=joe, lastName=smith, attributes={\"attr\":[\"1\",\"2\",\"3\",\"4\",\"5\"]}}"
-                + "]}}";
-
-        //when
-        Object result = executor.execute(query).getData();
-
-        // then
-        assertThat(result.toString()).isEqualTo(expected);
-    }     
-    
-    @Test
-    public void queryJsonEntityWhereSearchCriteria() {
-        //given
-        String query = "query {" + 
-                "  JsonEntities(where: {" 
-                +     "attributes: {LOCATE: \"key\"}" 
-                + "}) {" + 
-                "    select {" + 
-                "      id" + 
-                "      firstName" + 
-                "      lastName" + 
-                "      attributes" + 
-                "    }" + 
-                "  }" + 
-                "}";
-        
-        String expected = "{JsonEntities={select=["
-                + "{id=1, firstName=john, lastName=doe, attributes={\"attr\":{\"key\":[\"1\",\"2\",\"3\",\"4\",\"5\"]}}}"
-                + "]}}";
-
-        //when
-        Object result = executor.execute(query).getData();
-
-        // then
-        assertThat(result.toString()).isEqualTo(expected);
-    }
-    
-    @Test
-    public void queryTaskVariablesWhereSearchCriteria() {
-        //given
-        String query = "query {" + 
-                "  TaskVariables(where: {" 
-                +     "value: {LOCATE: \"true\"}" 
-                + "}) {" + 
-                "    select {" + 
-                "      id" + 
-                "      name" + 
-                "      value" + 
-                "    }" + 
-                "  }" + 
-                "}";
-        
-        String expected = "{TaskVariables={select=[{id=2, name=variable2, value=true}]}}";
-
-        //when
-        Object result = executor.execute(query).getData();
-
-        // then
-        assertThat(result.toString()).isEqualTo(expected);
-    }
-    
-    @Test
-    public void queryTaskVariablesWhereSearchCriteriaVariableBinding() {
-        //given
-        String query = "query($value: VariableValue!) {" + 
-                "  TaskVariables(where: {" 
-                +     "value: {LOCATE: $value }" 
-                + "}) {" + 
-                "    select {" + 
-                "      id" + 
-                "      name" + 
-                "      value" + 
-                "    }" + 
-                "  }" + 
-                "}";
-        
-        Map<String,Object> variables = Collections.singletonMap("value", true);
-        
-        String expected = "{TaskVariables={select=[{id=2, name=variable2, value=true}]}}";
-
-        //when
-        Object result = executor.execute(query, variables).getData();
-
-        // then
-        assertThat(result.toString()).isEqualTo(expected);
-    }
-    
-
-    @Test
-    public void queryProcessVariablesWhereSearchCriteriaVariableBindings() {
-        //given
-        String query = "query($value: VariableValue!)  {" + 
-                " ProcessVariables(where: {" 
-                +     "value: {LOCATE: $value}" 
-                + "}) {" + 
-                "    select {" + 
-                "      id" + 
-                "      name" + 
-                "      value" + 
-                "    }" + 
-                "  }" + 
-                "}";
-
-        Map<String,Object> variables = Collections.singletonMap("value", "[\"1\",\"2\",\"3\",\"4\",\"5\"]");
-        
-        String expected = "{ProcessVariables={select=[{id=1, name=document, value={key=[1, 2, 3, 4, 5]}}]}}";
-
-        //when
-        Object result = executor.execute(query, variables).getData();
-
-        // then
-        assertThat(result.toString()).isEqualTo(expected);
-    }
-
-    @Test
-    public void queryProcessVariablesWhereSearchCriteria() {
-        //given
-        String query = "query {" + 
-                " ProcessVariables(where: {" 
-                +     "value: {LOCATE: \"[\\\"1\\\",\\\"2\\\",\\\"3\\\",\\\"4\\\",\\\"5\\\"]\"}" 
-                + "}) {" + 
-                "    select {" + 
-                "      id" + 
-                "      name" + 
-                "      value" + 
-                "    }" + 
-                "  }" + 
-                "}";
-        
-        String expected = "{ProcessVariables={select=[{id=1, name=document, value={key=[1, 2, 3, 4, 5]}}]}}";
-
-        //when
-        Object result = executor.execute(query).getData();
-
-        // then
-        assertThat(result.toString()).isEqualTo(expected);
-    }
-
-    @Test
-    public void queryProcessVariablesWhereWithEQStringSearchCriteria() {
-        //given
-        String query = "query {" + 
-                " TaskVariables(where: {" 
-                +     "name: {EQ: \"variable1\"}" 
-                +     "value: {EQ: \"data\"}" 
-                + "}) {" + 
-                "    select {" + 
-                "      name" + 
-                "      value" + 
-                "    }" + 
-                "  }" + 
-                "}";
-        
-        String expected = "{TaskVariables={select=[{name=variable1, value=data}]}}";
-
-        //when
-        Object result = executor.execute(query).getData();
-
-        // then
-        assertThat(result.toString()).isEqualTo(expected);
-    }    
-    
-    @Test
-    public void queryProcessVariablesWhereWithEQBooleanSearchCriteria() {
-        //given
-        String query = "query {" + 
-                " TaskVariables(where: {" 
-                +     "name: {EQ: \"variable2\"}" 
-                +     "value: {EQ: true}" 
-                + "}) {" + 
-                "    select {" + 
-                "      name" + 
-                "      value" + 
-                "    }" + 
-                "  }" + 
-                "}";
-        
-        String expected = "{TaskVariables={select=[{name=variable2, value=true}]}}";
-
-        //when
-        Object result = executor.execute(query).getData();
-
-        // then
-        assertThat(result.toString()).isEqualTo(expected);
-    }    
- 
-    @Test
-    public void queryProcessVariablesWhereWithEQNullSearchCriteria() {
-        //given
-        String query = "query {" + 
-                " TaskVariables(where: {" 
-                +     "name: {EQ: \"variable3\"}" 
-                +     "value: {EQ: null}" 
-                + "}) {" + 
-                "    select {" + 
-                "      name" + 
-                "      value" + 
-                "    }" + 
-                "  }" + 
-                "}";
-        
-        String expected = "{TaskVariables={select=[{name=variable3, value=null}]}}";
-
-        //when
-        Object result = executor.execute(query).getData();
-
-        // then
-        assertThat(result.toString()).isEqualTo(expected);
-    }    
-
-    @Test
-    public void queryProcessVariablesWhereWithEQIntSearchCriteria() {
-        //given
-        String query = "query {" + 
-                " TaskVariables(where: {" 
-                +     "name: {EQ: \"variable6\"}" 
-                +     "value: {EQ: 12345}" 
-                + "}) {" + 
-                "    select {" + 
-                "      name" + 
-                "      value" + 
-                "    }" + 
-                "  }" + 
-                "}";
-        
-        String expected = "{TaskVariables={select=[{name=variable6, value=12345}]}}";
-
-        //when
-        Object result = executor.execute(query).getData();
-
-        // then
-        assertThat(result.toString()).isEqualTo(expected);
-    }
-    
-    @Test
-    public void queryProcessVariablesWhereWithEQDoubleSearchCriteria() {
-        //given
-        String query = "query {" + 
-                " TaskVariables(where: {" 
-                +     "name: {EQ: \"variable5\"}" 
-                +     "value: {EQ: 1.2345}" 
-                + "}) {" + 
-                "    select {" + 
-                "      name" + 
-                "      value" + 
-                "    }" + 
-                "  }" + 
-                "}";
-        
-        String expected = "{TaskVariables={select=[{name=variable5, value=1.2345}]}}";
-
-        //when
-        Object result = executor.execute(query).getData();
-
-        // then
-        assertThat(result.toString()).isEqualTo(expected);
-    }    
-    
-    
-    @Test
-    public void queryProcessVariablesWhereWithINSingleValueSearchCriteria() {
-        //given
-        String query = "query {" + 
-                " TaskVariables(where: {" 
-                +     "value: {IN: 1.2345}" 
-                + "}) {" + 
-                "    select {" + 
-                "      name" + 
-                "      value" + 
-                "    }" + 
-                "  }" + 
-                "}";
-        
-        String expected = "{TaskVariables={select=[{name=variable5, value=1.2345}]}}";
-
-        //when
-        Object result = executor.execute(query).getData();
-
-        // then
-        assertThat(result.toString()).isEqualTo(expected);
-    }    
-        
-    @Test
-    public void queryProcessVariablesWhereWithINListValueSearchCriteria() {
-        //given
-        String query = "query {" + 
-                " TaskVariables(where: {" 
-                +     "value: {IN: [1.2345]}" 
-                + "}) {" + 
-                "    select {" + 
-                "      name" + 
-                "      value" + 
-                "    }" + 
-                "  }" + 
-                "}";
-        
-        String expected = "{TaskVariables={select=[{name=variable5, value=1.2345}]}}";
-
-        //when
-        Object result = executor.execute(query).getData();
-
-        // then
-        assertThat(result.toString()).isEqualTo(expected);
-    }        
-    
-    @Test
-    public void queryProcessVariablesWhereWithNINListValueSearchCriteria() {
-        //given
-        String query = "query {" + 
-                " TaskVariables(where: {" 
-                +     "value: {NIN: [null]}" 
-                + "}) {" + 
-                "    select {" + 
-                "      name" + 
-                "      value" + 
-                "    }" + 
-                "  }" + 
-                "}";
-        
-        String expected = "{TaskVariables={select=["
-                + "{name=variable1, value=data}, "
-                + "{name=variable2, value=true}, "
-                + "{name=variable4, value={key=data}}, "
-                + "{name=variable5, value=1.2345}, "
-                + "{name=variable6, value=12345}, "
-                + "{name=variable7, value=[1, 2, 3, 4, 5]}]}}";
-
-        //when
-        Object result = executor.execute(query).getData();
-
-        // then
-        assertThat(result.toString()).isEqualTo(expected);
-    }
-    
-    @Test
-    public void queryProcessVariablesWhereWithNINSingleValueSearchCriteria() {
-        //given
-        String query = "query {" + 
-                " TaskVariables(where: {" 
-                +     "value: {NIN: null}" 
-                + "}) {" + 
-                "    select {" + 
-                "      name" + 
-                "      value" + 
-                "    }" + 
-                "  }" + 
-                "}";
-        
-        String expected = "{TaskVariables={select=["
-                + "{name=variable1, value=data}, "
-                + "{name=variable2, value=true}, "
-                + "{name=variable4, value={key=data}}, "
-                + "{name=variable5, value=1.2345}, "
-                + "{name=variable6, value=12345}, "
-                + "{name=variable7, value=[1, 2, 3, 4, 5]}]}}";
-
-        //when
-        Object result = executor.execute(query).getData();
-
-        // then
-        assertThat(result.toString()).isEqualTo(expected);
-    }
-    
-    @Test
-    public void queryProcessVariablesWhereWithNEValueSearchCriteria() {
-        //given
-        String query = "query {" + 
-                " TaskVariables(where: {" 
-                +     "value: {NE: null}" 
-                + "}) {" + 
-                "    select {" + 
-                "      name" + 
-                "      value" + 
-                "    }" + 
-                "  }" + 
-                "}";
-        
-        String expected = "{TaskVariables={select=["
-                + "{name=variable1, value=data}, "
-                + "{name=variable2, value=true}, "
-                + "{name=variable4, value={key=data}}, "
-                + "{name=variable5, value=1.2345}, "
-                + "{name=variable6, value=12345}, "
-                + "{name=variable7, value=[1, 2, 3, 4, 5]}]}}";
-
-        //when
-        Object result = executor.execute(query).getData();
-
-        // then
-        assertThat(result.toString()).isEqualTo(expected);
-    }        
-     
-    @Test
-    public void queryProcessVariablesWhereWithINListTypedValueSearchCriteria() {
-        //given
-        String query = "query {" + 
-                " TaskVariables(where: {" 
-                +     "value: {IN: [null, true, \"data\", 12345, 1.2345]}" 
-                + "}) {" + 
-                "    select {" + 
-                "      name" + 
-                "      value" + 
-                "    }" + 
-                "  }" + 
-                "}";
-        
-        String expected = "{TaskVariables={select=["
-                + "{name=variable1, value=data}, "
-                + "{name=variable2, value=true}, "
-                + "{name=variable3, value=null}, "
-                + "{name=variable5, value=1.2345}, "
-                + "{name=variable6, value=12345}"
-                + "]}}";
-
-        //when
-        Object result = executor.execute(query).getData();
-
-        // then
-        assertThat(result.toString()).isEqualTo(expected);
-    }     
-    
-    
-    @Test
-    public void queryTasksVariablesWhereWithExplicitANDEXISTSByNameAndValueCriteria() {
-        //given
-        String query = "query {" + 
-                "  Tasks(where: {" + 
-                "    status: {EQ: COMPLETED}" + 
-                "    AND: [{" + 
-                "      EXISTS: {" + 
-                "        variables: {" + 
-                "          name: {EQ: \"variable1\"}" + 
-                "          value: {EQ: \"data\"}" + 
-                "        }" + 
-                "      }" + 
-                "    } {" + 
-                "      EXISTS: {" + 
-                "        variables: {" + 
-                "          name: {EQ: \"variable2\"}" + 
-                "          value: {EQ: true}" + 
-                "        }" + 
-                "      }" + 
-                "    }]" + 
-                "  }) {" + 
-                "    select {" + 
-                "      id" + 
-                "      status" + 
-                "      variables {" + 
-                "        name" + 
-                "        value" + 
-                "      }" + 
-                "    }" + 
-                "  }" + 
-                "}";
-        
-        String expected = "{Tasks={select=["
-                + "{id=1, status=COMPLETED, variables=["
-                +   "{name=variable2, value=true}, "
-                +   "{name=variable1, value=data}]}"
-                + "]}}";
-
-        //when
-        Object result = executor.execute(query).getData();
-
-        // then
-        assertThat(result.toString()).isEqualTo(expected);
-    }       
-
-
-    @Test
-    public void queryTasksVariablesWhereWithEXISTSByNameAndValueCriteria() {
-        //given
-        String query = "query {" + 
-                "  Tasks(where: {" + 
-                "    status: {EQ: COMPLETED}" + 
-                "    EXISTS: [{" + 
-                "      variables: {" + 
-                "        name: {EQ: \"variable1\"}" + 
-                "        value: {EQ: \"data\"}" + 
-                "      }" + 
-                "    } {" + 
-                "      variables: {" + 
-                "        name: {EQ: \"variable2\"}" + 
-                "        value: {EQ: true}" + 
-                "      }" + 
-                "    }]" + 
-                "  }) {" + 
-                "    select {" + 
-                "      id" + 
-                "      status" + 
-                "      variables {" + 
-                "        name" + 
-                "        value" + 
-                "      }" + 
-                "    }" + 
-                "  }" + 
-                "}";
-        
-        String expected = "{Tasks={select=["
-                + "{id=1, status=COMPLETED, variables=["
-                +   "{name=variable2, value=true}, "
-                +   "{name=variable1, value=data}]}"
-                + "]}}";
-
-        //when
-        Object result = executor.execute(query).getData();
-
-        // then
-        assertThat(result.toString()).isEqualTo(expected);
-    }       
-
-
 }
